// src/app/module/subscription/subscription.interface.ts
import { SubscriptionStatus } from "@prisma/client";

export interface ISubscription {
  id?: string;
  tenantId: string;
  planId: string;
  stripeSubId: string;
  status?: SubscriptionStatus;
  stripeCustomerId: string;
  startDate: Date;
  endDate: Date;
  intervalCount?: number;
  createdAt?: Date;
  updatedAt?: Date;
}